:sparkles: feat(rag-service): Pattern analysis improvements & refactoring
